Add AttrTypeMap as serializing workaround
Currently, if we try to call `.json()` on any Pydantic model that
has `AttrType` field in it, Pydantic won't be able to convert it.
Pydantic allows custom serializers for type, so we can overcome this
issue for Pydantic. However, it seems that FastAPI ignores custom
serializers and breaks when it comes across our custom types.
That's why I wrote this workaround. There will be no need in it when
this (or similar) PR gets merged to FastAPI repo:
https://github.com/tiangolo/fastapi/pull/2061