Remove custom validator from AttributeCreateSchema
It seems that our validator is conflicting with FastAPI/Pydantic
behavior and when we pass type as string it doesn't convert it to enum
value as it should do by default. I could add another elif for handling
strings, but I think that we won't lose much if we won't be able to pass
AttrType objects to AttributeCreateSchema, so I removed custom validator
completely. Now it's supposed to accept `str` and `AttrTypeMapping`
values for `type` field