Set restrictions on attribute names
Now attribute names can only be valid python identifiers (1) that don't
start(2)/end(3) with underscore

Why (1): because to build API docs there is a place where we make a
dataclass with fields equal to attribute names

Why (2): because Pydantic may ignore fields starting with an underscore

Why (3): to avoid shadowing when we create an attribute with name equal
to member of pydantic.BaseModel we add `__` to the fieldname and alias
with actual fieldname to make it possible to create such schema. So, if
we try to create schema with fields `shadowed` and `shadowed__`, when
generating docs there will be only `shadowed__` which corresponds either
to initially passed `shadowed` or `shadowed__` field. To avoid such
problems we should just forbid attrs ending with underscores

Why (2,3): another reason to forbid underscores is shadowing dunder
methods