Fixed issue with incorrect 404 message when following foreign key link
When following a link to an entity, it can happen that the API request to get the entity details
is sent before the injected property `activeSchema` is updated. This means that the request used
the wrong URL.

With this fix, the schema slug is used directly from the current route parameters.

Fixes #158.