Bugfix for `authorized_user`
The query to obtain the specific entity ID based on the slug used in the request URL did not take
into account that another entity in a different schema is allowed to have the same slug. This could
result in a `MultipleResultsFound` exception.

This change limits the query to the schema defined in the request URL and avoids the exception.

In addition, the extraction of the schema slug from the URL was made stricter.

Closes #165.